requirements.txt

fastapi==0.110.0
uvicorn==0.29.0
qdrant-client==1.7.3
sentence-transformers==2.5.1
openai>=1.30.0
numpy==1.26.4
pydantic==2.7.1
httpx==0.27.0
python-dotenv==1.0.1
